# backend/openapi.yaml
openapi: 3.1.0
info:
  title: SMKN 1 Jakarta School Profile CMS API
  version: 1.0.0
  description: Standalone Express API for the public school profile CMS.
servers:
  - url: http://localhost:4000
security: []
tags:
  - name: Public content
  - name: Chat
  - name: CMS
paths:
  /health:
    get:
      tags: [Public content]
      operationId: health
      responses:
        "200":
          description: Service is responding.
  /api/public/{resource}:
    get:
      tags: [Public content]
      operationId: listPublishedContent
      parameters:
        - $ref: "#/components/parameters/Resource"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Published, currently visible content.
        "400":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/public/{resource}/{id}:
    get:
      tags: [Public content]
      operationId: getPublishedContent
      parameters:
        - $ref: "#/components/parameters/Resource"
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Published item.
        "404":
          $ref: "#/components/responses/NotFound"
  /api/public/chat:
    post:
      tags: [Chat]
      operationId: schoolProfileChat
      description: OpenAI-compatible chatbot endpoint. Logs category, outcome, model, and timestamp only; no question, IP, or conversation is persisted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [message]
              properties:
                message:
                  type: string
                  minLength: 1
                  maxLength: 2000
      responses:
        "200":
          description: Chat answer.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    required: [reply, model]
                    properties:
                      reply: { type: string }
                      model: { type: string }
        "400":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          description: Chat provider is unavailable.
        "503":
          description: Chat provider is not configured.
        "504":
          description: Chat provider timed out.
  /api/admin/{resource}:
    get:
      tags: [CMS]
      operationId: listAdminContent
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Resource"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200": { description: Draft and published content. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [CMS]
      operationId: createContent
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Resource"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContentInput" }
      responses:
        "201": { description: Content created. }
        "400": { $ref: "#/components/responses/InvalidInput" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/me:
    get:
      tags: [CMS]
      operationId: getCurrentCmsUser
      description: Returns the authenticated user's server-verified CMS role.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Authenticated user and assigned CMS role.
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/analytics:
    get:
      tags: [CMS]
      operationId: getChatbotAnalytics
      description: Returns category and outcome aggregates for the last 30 days, plus recent metadata only. No prompts, replies, user identifiers, or IP addresses are returned.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Privacy-safe chatbot analytics.
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/{resource}/{id}:
    patch:
      tags: [CMS]
      operationId: updateContent
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Resource"
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContentInput" }
      responses:
        "200": { description: Content updated. }
        "400": { $ref: "#/components/responses/InvalidInput" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [CMS]
      operationId: deleteContent
      description: Admin-only hard delete; editors may create and update but cannot delete.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Resource"
        - $ref: "#/components/parameters/Id"
      responses:
        "204": { description: Content deleted. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/users/{userId}/role:
    put:
      tags: [CMS]
      operationId: assignCmsRole
      description: Admin-only. The target must already exist in Supabase Auth; client metadata is never used to authorize requests.
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [role]
              properties:
                role:
                  type: string
                  enum: [admin_it, editor_guru_staf]
      responses:
        "204": { description: Role assigned. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { description: The acting administrator attempted to change their own role. }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Supabase access token
  parameters:
    Resource:
      name: resource
      in: path
      required: true
      schema:
        type: string
        enum: [news, announcements, academic-agenda, hero-banners]
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, maximum: 100000, default: 0 }
  schemas:
    ContentInput:
      description: Resource-specific strict object; see backend/src/service.mjs for exact per-resource field limits.
      type: object
      additionalProperties: false
      properties:
        title: { type: string, maxLength: 200 }
        slug: { type: string, maxLength: 160 }
        body: { type: string, maxLength: 50000 }
        excerpt: { type: [string, "null"], maxLength: 500 }
        description: { type: [string, "null"], maxLength: 5000 }
        image_url: { type: string, format: uri }
        cover_image_url: { type: [string, "null"], format: uri }
        link_url: { type: [string, "null"], format: uri }
        is_published: { type: boolean, default: false }
        published_at: { type: [string, "null"], format: date-time }
        starts_at: { type: [string, "null"], format: date-time }
        ends_at: { type: [string, "null"], format: date-time }
        severity: { type: string, enum: [info, important, urgent] }
        display_order: { type: integer, minimum: 0, maximum: 10000 }
  responses:
    InvalidInput:
      description: Request did not satisfy validation rules.
    Unauthenticated:
      description: A valid Supabase bearer token is required.
    Forbidden:
      description: The verified Supabase user does not have the required CMS role.
    NotFound:
      description: Resource not found or not publicly visible.
    RateLimited:
      description: Request rate limit exceeded.